Implement collation_check_walker()
collation_check_walker() walks the query parse tree to detect nodes that
have a non default collation (DEFAULT_COLLATION_OID for collatable types
and InvalidOid otherwise). If any such node is detected, we immediately
fall back to planner. For this reason, it is called early in the Query
to DXL translation process.

Signed-off-by: Melanie Plageman <melanieplageman@pivotal.io>