ExecHashJoin: allow null values in hash tables for nonequijoin
For IS NOT DISTINCT joins, GPDB has its own concept of whether a hash
table should retain tuples that have null hash keys: HashState.hs_keepnull,
the use of which was inadvertently removed from the code during a
conflict resolution in this iteration. Upstream added the concept of
"keep nulls" to ExecHashTableCreate() in 9.1, so make use of that same
parameter here.

This fixes the qp_union_intersect failure in ORCA, which was caused by
joins losing rows with null hash keys.

Co-authored-by: Taylor Vesely <tvesely@pivotal.io>