Cast a int16 variable to int to align with the %d in format string in FtsTestSegmentDBIsDown()
We saw a panic (sigsegv) at FtsTestSegmentDBIsDown()->ereport(), however we could not further debug
since the environment was gone very soon. Since there is a lot of format conversion in ereport
so we can not guess which conversion causes the crash. Although we suspect that here our change
is not the culprit, it is a good habit to make format and variable align.

Here is the panic stack.

0x00007f33b912e94f in __strlen_sse42 () from /lib64/libc.so.6
0x00000000009f7574 in write_message_to_server_log ()
0x00000000009fbee0 in send_message_to_server_log ()
0x00000000009ff865 in EmitErrorReport ()
0x00000000009fc9d5 in errfinish ()
0x0000000000acd6cf in FtsTestSegmentDBIsDown ()
0x0000000000abd3bb in createGang_async ()

Co-authored-by: Jinbao Chen <jinchen@pivotal.io>