WIP: check for relfile collision and error out
The disappearing catalog issue appears to be caused by temporary
relations' shared buffers colliding with those owned by permanent
relations. As a stop-gap, check for temp relfilenode collision during
GetNewRelFileNode().

Temporarily elog(ERROR) when a collision happens to prove that this is
in fact the issue we're seeing. Assuming this works as expected, we'll
remove the error later.