get_variable_range: ERROR instead of crashing with broken stats
Commit 88b2ab36 from gpdb/master brought in additional tests for broken
histogram data when performing cost analysis during planning. This
crashes the iteration_8 mergejoin planning code, because it didn't check
for the type mismatch.

Since we already perform this check for MCV mismatch later, do the same
for histogram data, and add an explicit test to catch this case. (The
additional test is less useful for ORCA; oh well.)

This is the second time we've had to fix something like this, so a FIXME
has been added to revisit.

Co-authored-by: Asim Praveen <apraveen@pivotal.io>