Fix pg_proc catalog in pg_proc_gp.h by fixing catullus.pl and regenerating.
pg_proc adds proleakproof and protransform in pg9.2.

We now allow the leakproof attribute in pg_proc.sql to set proleakproof.

For protransform, since this is not widely used and is a trival thing, we
hardcode as 0 at this moment and add a FIXME in case in the future
some new functions want this.