Fixed the sync rep hang issue.
The root cause of the sync rep hang is that standby priority
is set to zero, which means 'not-listed'. Hence, when wal sender
receives reply message, it fails to wake up waiters as it is not
regarded as a potential sync standby.

This commit fixed the hang by always setting standby priority to 1.
GPDB currently does not apply the concept of standby priority, as
GPDB allows max 1 walsender to be alive at a time.