Fix issues in CTAS code.
Note single mode CTAS could work now (in utility mode) but mpp CTAS is failing.

In this patch, I added a commmon interface rInitPlan() in struct DestReceiver
so that the code that creates relation on cluster in CTAS runs in that
function, which is called in InitPlan(), instread of in rStartup() of DR_intorel.

I could of course add the interface in DR_intorel only and hack to
call it for CTAS in InitPlan() (Refer OpenIntoRel in gpdb master),
but that seems to be uglier.

MPP CATS code cdbparallelize() and apply_motion() are obviously buggy.
And note that now we plan the query in CTAS separately so we need to
enforce it as MPP.