Remove unnecessary assert in SessionState_Acquire
There was defensive assertion in SessionState_Acquire that defends against a
race condition that should never be possible, because everything that modifies
the AllSessionStateEntries->freeList is protected with an exclusive lock.

After the merge with postgres 9.2, the ExceptionalCondition call from an Assert
is quilified with __attribute__((noreturn)). Because the test was
(intentionally) triggering and then ignoring an Assertion, after setting the
noreturn attribute, the compiler no longer expects us to be able to return from
an Assert at all. The result is that the compiler no longer expects that we can
return from the assertion, and when we do, the unit tests exhibits undefined
behavior.

Co-authored-by: Jesse Zhang <sbjesse@gmail.com>