Fix test failiures in select_views and function_extensions.
There should be no failure in test select_views now. Fro the case in
theory the NOTICE output could vary in order because of the different
respsonse speed of segments, however QD seems to send query to
segments in sequence so I'm assuming that mostly the output should
be in sequence normally. In the future if we see the disorder of
NOTICE output we should modify our perl script to order the output.

function_extensions is partially fixed. There is a planner cost issue
that needs some additional effort.