Need to recalculate initial cost before calculating the final cost of join paths.
Outer path and inner path for join could be modified during mpp-planning.
We need to recalculate the initial cost for those join paths when calculating
the final costs (e.g. before final_cost_nestloop()), else the costs will be
wrong and the plan could be wrong.

This fixes the test failure of function_extensions.