Fix drop inherit table failure
create table t(c1 int);
create table th(c1 int) inherits (t);
drop table t, th;

The drop statement works on master but fails on segments, which is
not the same as upstream.
In GPDB, master can find 'th' depends on 't' so it deletes the two relations
in performMultipleDeletions, but it dispatches two different drop commands
for table 't' and 'th' to segment, and segment can't drop 't' because it has
depend object.