Revert "Fixed select_into."
This reverts commit 09509042421ad8205b47c85d950b766021becab9.

Previously we create the table b since the table is
created in the inherit test (like in pg) however that
case is disabled for a long time on gpdb. Now that we
just enabled the inherit test, we should remove the table
create sql although it is really not a good experience
of using a common table name 'b' from another case.