Set rows variable correctly for some code path.
It seems that cdbpath_rows() should be used for cost estimation only
however it is often called to set rows also. We should be very
careful on this. In the long run we should decouple cost estimation
and rows variable setting. That is to say, remove the ugly hack function
cdbpath_rows() and its use.