Fix gpdiffcheck test failure.
pg 9.2 changes the logic of path pruning and selection (see add_path())
so sub-optimal results could be encountered easiler. In this test case,
if we disable that functionality in code by hacking, the plans will be
same as those in gpdb master. I'm changing the exp output to follow current
planner output after some analysis. Most importantly this case
is used to test the gpdiff script (so we see complex plan) so we should not
spend too much on planner itself in this test case.