Fix GDD under fastpath.
With 9.2 merge fastpath is add, GDD is disable temporarily as it cannot properly
recognize all locks with fastpath.

In the function set the HoldTillEndXact field, locallock->lock might be NULL if
the lock is held via fast path.

In this commit, I add a field fpHoldTillEndXactBits(like fpLockBits) in PROC
struct to save this info.