Change how we implement pg_locks
As in Postgres, the view pg_lock in Greenplum is implemented by a
function pg_lock_status.
Unlike in upstream Postgres, Greenplum's pg_lock_status aggregates locks
from all the primary segments plus the master. To do that, we dispatch a
SQL string to all the segments, and do the "concatenation" on master.
This simplistic approach has served us long (since at least the 4.x
release) and well, but it is also flawed: it only works for simple
queries like `SELECT * FROM pg_locks`. As soon as you join the view with
a distributed table, it will either return the wrong results (only
showing locks from master), or yield plans that are not even executable
-- "query plan with multiple segworker groups is not supported" or
"writer gang is busy".

The root problem with that approach is hiding a dispatch inside of a
function from an optimizer (planner / ORCA). Without that knowledge, no
optimizer is safe from making a plan that is un-executable.

This commit rewrote the pg_locks view into a UNION ALL between two
set-returning functions gp_lock_status_master() and
gp_lock_status_segments() with appropriate `EXECUTE ON` attributes.
Underlying both is the same pg_lock_status function. This eliminates the
need to dispatch from within pg_lock_status. In fact, we were able to
vastly reduce the diff with upstream in lockfuncs.c .

This is discovered while working on the isolation2 test case
"starve_case".
Taylor taught us how to properly modify the catalog, and provided lots
of feedback in the design.

Co-authored-by: Jesse Zhang <sbjesse@gmail.com>
Co-authored-by: Alexandra Wang <lewang@pivotal.io>
Co-authored-by: Taylor Veseley <tveseley@pivotal.io>