Fetch pg_upgrade from master.
* Break out Greenplum specific parts from pg_upgrade

This attempts to minimize the diff to upstream in preparation of
merging with a later version of pg_upgrade. In some cases this
re-introduces quirks from small style cleanups we've performed,
but it will only be for a short period until the merge happens.

PostgreSQL have in recent versions moved to just having a version.c
file for version specific function. In this commit we move Greenplum
specific code to version_gp.c, since we would do that sooner or later
anyways.

No functionality is altered with this, this is limited to refactoring.

* Import pg_upgrade 9.3.22 sources

This overwrites the Greenplum pg_upgrade version with the upstream
9.3.22 sources. No attempt is made to make this compile or at all
be even remotely useable. Follow-up commits will re-introduce all
functionality which was lost.

* Reapply the Greenplum specific code on top of pg_upgrade

The previous commit backported the PostgreSQL 9.3 pg_upgrade code
base and overwrote the hacks to allow for upgrading a Greenplum
cluster. This commit brings back the patchsets that were lost, and
adapts them to the new pg_upgrade code as well as mildly refactors
them to take advantage of new opportunities.

* Replace Oid preallocation with pg_upgrade Oid handling

Previously, the Greenplum pg_upgrade was syncronizing the Oids across
upgrades using the backend Oid pre-allocation. All pre-allocations
were generated in an Oid file, which was executed before the schema
was restored. When doing parallel restores, this strategy becomes
problematic as the Oid preallocation may be done in a backend other
than the one where the Oid is required. This commit rips out the use
of Oid preallocation in favor of using the upstream Oid handling in
pg_upgrade where the Oids are set just before being required. A major
difference is that we use the Oid preassignment scaffolding under the
covers, but in a more upstream merge friendly way.

For now, a known regression is that toast tables which in the old
cluster doesn't match the naming convention in the new cluster fails
to restore. Drop before running test until we know how to handle these.

* Remove hard-coded "\connect postgres" from pg_dumpall.

This doesn't appear to accompish anything useful, and does make the
restore fail if the postgres database happens to have been dropped.

* First stab of changing pg_upgrade distribution method

Previously each QE would dump/restore the schema before upgrading
the datafiles, which consumes a lot of time on large databases.
Instead, use the datadir which was created from the dump/restore
on the QD for the QEs as well and bootstrap the segment upgrade
to jump straight to copying files instead.

This is a first stab at implementing this model of distribution,
follow-up commits will be required to finalize the patch.

* In pg_upgrade, fix bug where no users were dumped in pg_dumpall
binary-upgrade mode;  instead only skip dumping the current user.

This bug was introduced in during the removal of split_old_dump().  Bug
discovered during local testing.

* Fix vacuum check in pg_upgrade testing

Make sure we use the psql client from the new bindir (the on in PATH
might well be an upstream postgres psql etc), and use a port different
from the standard postgres port since it's likely to be in use.

Also fix integer value comparison which caused a warning in my Bash,
while it might work in other versions of Bash.

* Ensure we have a regression.diffs file to satisfy the pipeline

This is extracted from a larger patch in 5e92c436adcb32d295e77b3f3.

Co-authored-by: Paul Guo <paulguo@gmail.com>
Co-authored-by: Max Yang <myang@pivotal.io>

* Temporarily minimize the pg_upgrade test dataset

The regression database has shown to be full of questionable
objects which we probably won't support for upgrades without
manual intervention before the upgrade. With all the changes
going into pg_upgrade it's time to start on a known minimal
dataset and expand from there to the full ICW leftover schema.
As a stop-gap for now, drop the regression databases eand only
upgrade a small subset.

Also disallow upgrades of orhpaned toast tables, and clean
up the invocation of the test_gpdb_pre.sql script to be less
hacky.

* Fix DDL syntax in pg_upgrade test

* Update kernel.sem setting to be consistent with 5.9.0 doc (#5333)

Between GPDB 5.8.0 and 5.9.0, the recommended Linux System Settings
changed for kernel.sem.
https://gpdb.docs.pivotal.io/580/install_guide/prep_os_install_gpdb.html#topic3
https://gpdb.docs.pivotal.io/590/install_guide/prep_os_install_gpdb.html#topic3

However, the gpdb repository had references to the kernel.sem parameter,
as well as a couple other parameters, which seem to be inconsistent.

Co-authored-by: Amil Khanzada <akhanzada@pivotal.io>
Co-authored-by: Trevor Yacovone <tyacovone@pivotal.io>

* Move hyperloglog from contrib to catalog

This commit moves hyperloglog aggregates and functions to catalog as
well as migrates any hyperloglog related functions into
`src/backend/utils`.

* Fixes after sync with master

* Revise CTAS to behavior correctly when pg_upgrade