Set up a new session group in postmaster.
Without this patch, pressing ctrl+c to terminate regression tests
could always cause master processes down if master is identical as client.

This issue 100% happens after running into tests in greenplum_schedule.
It is because the first test case instr_in_shmem_setup in
greenplum_schedule restarts the cluster, however without this patch,
when we press ctrl+c to terminate testing, SIGINT will
be delivered to postmaster process also.

This issue does not happen on gpdb master at this moment due to
the existence of pmdaemonize() which was removed in PG 9.2, and
since the logic on PG seems to be never changed since 9.2 it
happens on PG master also.

Not sure how PG upstream thinks about this, but I'm checking in
this patch at first since this really affect development.