Fix SIRV transformation of function returning composite types.
For example, imagine that you had a query like this:

SELECT address, name FROM test()

Previously, this was transformed into:

SELECT sirvf_sq.address, sirvf_sq.name FROM (SELECT test()) AS sirvf_sq

I.e. the Vars referencing the columns returned by test() function were
replaced with FieldSelect expressions, to pick the columns from the
single composite datum returned by the sirvf_sq subquery. However, that
transformation was not done correctly for possible WHERE clause in the
query, only for the target list. So if the query had a WHERE clause:

SELECT address, name, FROM test() WHERE name = 'foo'

You got an error:

ERROR:  subquery testfunc does not have attribute 2

The most straightforward fix would have been to also transform Vars in the
WHERE clause. However, I'm a bit worried that we'll have more bugs like
this in other clauses, even if we fix the WHERE case. We could perhaps call
query_tree_mutator() to find and replace all the Vars everywhere in the
query, but because these functions are modifying the query in place, that'd
also require some restructuring of the code.

So instead, I changed the approach so that we add yet another subquery, to
contain the FieldSelects, and leave the original Vars unmodified. The above
query is now turned into:

SELECT address, name FROM (
  SELECT (sirvf_sq.test).address, (sirvf_sq.test).name
  FROM (SELECT test()) AS sirvf_sq
) as test;

Fixes github issue #2865. A modified version of the repro from that issue
is added to the regression test suite.

NOTE: Cherry-picked this fro mhttps://github.com/greenplum-db/gpdb/pull/5477
If that ends up not benig committed in this form, need to revert/fix this!