Resolve conflicts in xlog.c
Tried best to resolve the conflicts in this file based on diff compared to
upsteam file and keeping gpdb changed where relevant.

The file doesn't compile yet and need to fix lot of errors and warnings to make
it work. Mostly in gpdb addition code relating to type change for `XLogRecPtr`
and removal of macros for the same.

One of the hurdle faced was upstream now has function name same as old gpdb
function `XLogReadRecord()`. Upstream static `ReadRecord()` function was exposed
in gpdb as `XLogReadRecord()`.  For now to avoid the confusion I have renamed it
to `GP_ReadRecord()`.

No doubt this was very messy work, so there could be mistakes in resolution. If
the resolutions doesn't sound correct or if there is easier, cleaner way, feel
free to revert this commit and start fresh.