Fix assertion failure, when there is no cheapest_startup_path.
This can happen for LATERAL queries. The query I used to debug this was this
from rangefuncs_cdb:

 -- supposed to fail with ERROR - requires LATERAL
 select * from foo2, foost(foo2.fooid) z where foo2.f2 = z.f2;

It still fails at execution time, though, because the planner produces a
plan with a Motion such that the parameter cannot be passed from the outer
side of the  nested loop join to the inner side. But that's a different
problem.