Fix more issues caused by 'zapping' rte->funcexpr in setrefs.c.
The 'funcexpr' is zapped from the RTE, so we won't find it in
addRemoteExecParamsToParamList() anymore. Instead of scanning the rtable,
fix the param_walker, which walks the Plan tree, so that it knows to look
into the TableFunctionScan.funcexpr field.

I was hunting the remaining issues in 'table_functions' test, but this also
explains the failure in 'insert' and a few other tests, with VALUES
expressions. The values_list in RTEs were also zapped.