Fix error code in errors related to COPY FREEZE.
This was fixed in upstream in commit 09cecdf285. It was only backpatched
to 9.5, on the grounds that changing error codes "doesn't seem like a good
idea", but we should be OK for GPDB 6. If we left these as internal errors,
you get the (copy.c:3381) suffix in the error message, and we don't want
to have that in expected output.