Add error position to error on coercing partition boundaries.
I think this is more clear. Now that we the error position is displayed
with the error, we don't need to work so hard to deparse the value and the
partition type. The error position makes that clear.

Fixes regression failure on 'partition' test.

NOTE: This could be cherry-picked into a separate PR.