Simplify 'partindex_test', by using OUT parameters and pg_get_expr().
OUT parameters make calling the function much less awkward, as you don't
need to specify the output columns in every invocation. Also, change the
datatype of a few columns to pg_node_tree, so that you can use pg_get_expr()
to pretty-print them.

I'm doing this to hide the trivial differences in the internal string
representation of expressions. Currently, the test is failing because
'funcformat' changed, and it will also break if any new fields are added
or renamed in FuncExpr or Const nodes. Use pg_get_expr() to print the
fields in human-readable format, which is not sensitive to small changes
like that, rather than in the low-level node tree representation.

NOTE: This could be cherry-picked to master.