Silence failure in PL/python test, caused by NOTICE about distribution key.
We hide these with the init_file normally, but because the CREATE TABLE
was issued inside a function, you also got CONTEXT lines for the NOTICE,
and those were not hidden by init_file. We might need to make the
init_file and/or gpdiff smarter, if we get more of this, but for now,
just adjust the expected output.