Fix EXPLAINs in qp_gist_indexes2 test.
gpdiff only recognized EXPLAINs as special, when the EXPLAIN and the SELECT
are on the same line. Because of that, the costs were included in the
expected output, and were compared. Thanks to small changes in cost
estimation in 9.3, many of the EXPLAIN tests started to fail.

To fix, start using EXPLAIN (COSTS OFF). Alternatively, we could've moved
the SELECTs to the same line as EXPLAINs, and let gpdiff mask out the
costs, but I like the COSTS OFF style more for these tests.