Remove mock test for cdbappendonlystorage.c
I moved the appendonly_redo() function from cdbappendonlystorage.c to
cdbappendonlyxlog.c, which broke the mock test. It's not straightforward
to fix, because the test relied on mocking some functions that are in
cdbappendonlyxlog.c, and now that the function-under-test is in the same
file, the default rules for what to mock and what to not mock don't work.

However, looking at the test, it seems pretty uninteresting. It basically
tests that there is call to IsStandbyMode() in appendonly_redo(), and that
the switch-case statement works. I don't think we need bespoken tests for
those things. So rather than work hard to fix the test, just remove it.

NOTE: this could be cherry-picked to a separate PR, to give some more
visibility to this decision.