Disable most isolation tests, because they don't work on GPDB, and never have.
As noted in the comment at the top of the file, most of these don't work
as intended on GPDB, because UPDATEs and DELETEs on GPDB lock the whole
table. Also, GPDB doesn't support true serializable mode.

I kept a few that I think should work on GPDB, too, but are currently
failing. They need further investigation.