Fix another expected output for EXPLAIN table alias change.
When I looked at this earlier, I noticed that the date value in the
Filter had changed. Now I noticed that it's because the test uses
CURRENT_DATE. So we're done here, just adjust the expected output. I
didn't commit the change to the Filter line, but if you run the query,
it will always display the current date there.