Change test function, so that planner chooses same plan as before.
This was a similar case as in qp_functions_in_subqueries earlier. Thanks
to upstream commit 31c7c642b6, I think, the planner chose a different plan
for this query, and with a different plan, the query fails in a different
way than before. Hint the planner that the SRF only returns a few rows, to
get the old plan back.