For AO avoid error "tuple to be updated was already modified"
Upstream added this check mostly for triggers case. In GPDB, for AO tables
HeapTupleSelfUpdated is returned only in case of same command tuple update based
on visimap dirty list checking. In context of this errors for triggers case
visimap update will not return HeapTupleSelfUpdated and hence will not flow
through this code path.

Also, the check is based on hufd. hufd is not initialized and can't be for AO
case, as visimap update within same command happens at end of command.