Silence failure in 'subselect' test.
Amazingly, this started failing now, when I moved a test query that
used to be before this query, to later in the file. (Or when I moved
this query to its current place, if you want to think of it that way.)
The reason is that the query that used to be just before this was failing
in GPDB (for legitimate reasons). Because it was throwing an error, rather
than returning a result set, gpdiff considered this query and the
preceding failing query as one query, and consider the ORDER BY in the
second query as dictating a result order. Now that the preceding failing
query was moved, these queries are considered on their own, and gpdiff
sees the ORDER BY, and thinks that the result must be ordered.

That's my theory anyway. I tested these in GDPB 5, and got exactly the
same plans, and nondeterministic row ordering, so in any case I believe
this is supposed to pass.

NOTE: Another candidate for cherry-picking to separate PR, along with
moving the test query to where it is in upstream.