Fix is_insertable_into column in information_schema.tables.
We lost the change in the view definition in the merge. It should call
the pg_relation_is_updatable() helper function now.

While we're at it, fix pg_relation_is_updatable() to only mark writable
external tables as insertable. And add a test for that.

This fixes some of the failures in 'updatable_views', but not all.