Fix expected output for different plans in GPDB vs upstream.
In addition to the extra Motion and SplitUpdate nodes, the planner chose
Seq Scans over Index Scans for many of the queries. That seems OK, the
point of these tests isn't what kind of a scan the planner chooses, but
rather that the update/delete happens on the correct base table.