Accept different error message in GPDB.
In GPDB, queries with DISTINCT are sometimes (always?) transformed into
GROUP BY queries, already at parse analysis. It might be a bit confusing
to users, that the error message says GROUP BY, when the original query
had DISTINCT in it. But if you look at the view definition with \d+,
it really has GROUP BY in it at this point.