Update expected output for corrected plan.
The old plan was, in fact, wrong. The "i = 1" condition must not
be pushed down below the window functions. The new plan is correct.

NOTE: Opened github issue https://github.com/greenplum-db/gpdb/issues/5762
to track this. The new code is OK, but we should perhaps fix this in
stable branches, too, and add a more explicit test case.