Fix merge conflicts in join.out.
In the upstream, the plans changed from Index Scans to Bitmap Index Scans,
thanks to some code changes in the planner to consider bitmap index scans.
But in GPDB, the planner has been choosing different plans for these all
along. Keep the old plans.

There are more failures in 'join' test, but this is a start.