Fall back to Postgres planner on queries with LATERAL joins.
ORCA can't handle them. On my laptop, with a debug build of ORCA, LATERAL
queries threw exceptions from inside ORCA or the ORCA translator, and fell
back, even without this patch. But I suspect these are causing the crashes
on the concourse pipeline. Better to fall back cleanly, in any case.

NOTE: This is a TODO for ORCA.