Attempt fixing "corrupted statistics file" warnings.
The heap-repeatable-read-vacuum-freeze isolation test was failing in
the concourse pipeline with:

WARNING:  corrupted statistics file "pg_stat_tmp/global.stat"

I didn't see this on my laptop, but I think it's because I don't run with
resource queues.

I believe it was caused by the missing handling of the GPDB-specific 'Q'
entries in the stats file, in the new pgstat_read_db_statsfile_timestamp()
function. The rest of the changes in this patch are cosmetic cleanups for
things I spotted while trying to find the issue.