Avoid gpinitsystem to incorrectly report exit status 1.
Gpinitsystem greps for logs and check if any warnings or errors are
generated. If any returns exit status as 1. This logic is super error prone. For
example in recent merge iteration started getting this new line of output during
initdb "LOG","00000","Computing Scalar Stats : column
logerrors",,,,,,"ANALYZE;". It was reporting error because of matching this
logerrors. For now adding just "LOG" messages to be skipped from this check.

Ideally, long term the script should rely on return codes from utilities it runs
instead of scanning log file for warnings or errors. But that's big tomorrow
thing to solve.