Refactor pg_regress searching of *_optimizer.out files.
Move the code to use the altarnative *_optimizer.out or *_resgroup.out
files from pg_regress_main.c, into pg_regress.c. The main benefit of this
is that we can use *_optimizer.out files also in the isolation tests (In
the isolation tester, isolation_main.c is used instead of pg_regress_main.c,
so if the logic was in pg_regress_main.c, we'd need to duplicate it in
isolation_main.c).