Don't require PL/perl support on Windows.
We only build the clients, so no need for libperl. We do still need Perl
to compile, because the many scripts in the build process are written in
Perl, but we don't need PL/perl. I'm hoping that this will fix the Windows
build failure on concourse.

NOTE: extract to separate PR if this works.