Bump catversion, and update MULTIXACT_FORMATCHANGE_CAT_VER to match.
Bumping the catversion wasn't strictly necessary, but seems nice to keep
it more close to the date that the merge commit is actually going to be
pushed on.

As pointed out by Jacob.

NOTE: Shouldn't the other *_CAT_VER constants be updated for GPDB as well?
Some of them had been changed, while others had the upstream value.