Remove ssl renegotiation support.
There were a bunch of conflicts in the code, but rather than fix them,
let's just remove it. SSL renegotiation support is being removed in PG 9.5
anyway, so we might as well lean a bit forward and cull it now, rather
than try to make it work.

This is a backport of 426746b930.