Avoid catalog cache lookups during abort processing.
SetResQueueId() is called when resetting current role GUC at abort. Not
cool to do catalog lookups at that time. Instead, just clear the current
cached queue id, and move the lookup to GetResQueueId(). Hopefully, all
calls of GetResQueueId() happen in a valid transaction.

NOTE: If this works, would be good candidate for extracting to separate PR.