Quick fix for hanging "SET SESSION AUTHORIZATION" commands.
SET SESSION AUTHORIZATION would call superuser(), which would do a
catalog lookup, which would try to get a distributed snapshot. In
a QE reader, it would hang, waiting for the QE writer to write the snapshot
to disk.

Not sure if there's some other way to fix this, but it seems reasonable
to always use local snapshots for catalog lookups.