Refactor the way Split Update nodes are constructed in the planner.
One specialty in a Split Update is that the node needs the *old* values
for all the updated rows, to compute the distribution hash for each
old row, so that they can be deleted. That was previously handled at
the time when the SplitUpdate node was created, by adding any missing
Vars for the old values to the subplan's target list, pushing them down
through joins and any other plan nodes, all the way down to the Scan
node for that relation. That seemed complicated and fragile.

The reason to tackle this right now is that we were seeing failures related
to this, while working on the PostgreSQL 9.4 merge. It added a test case,
where a Split Update was done through a security barrier view. The security
barrier view added a SubqueryScan to the plan tree, and the mechanism to
push through the old attributes couldn't cope with that. I'm sure we
could've hacked that to make it work, but this refactoring seems like a
better long term fix.

This patch makes it the responsibility of preprocess_targetlist(), to ensure
that the old values are made available to the top of the tree, if a Split
Update is needed. preprocess_targetlist() seems like the appropriate place,
because it already does that for columns that are not modified by the
UPDATE.

Now that we are making the decision on whether to do a split update in
preprocess_targetlist() already, add a flag to PlannerInfo to remember that
decision, to the point where the ModifyTable node is added to the top of
the plan tree.

Also add a test case, for an inherited table where some children have a
different distribution key, and an UPDATE on some of the children require a
Split Update, and others don't. That was causing me trouble at one point
during the development, and I'm not sure if there was any existing test to
cover that.

(See https://github.com/greenplum-db/gpdb/pull/5870)