Fix conflicts in pg_dump directory.
Resolved most of conflicts caused due to not having code for < 8.0. Maybe we
should keep all the code and avoid future merge conflicts.

Had to modify few queries to account for upstream modifications in 9.4. Tried
our best but testing would really reveal if they need to be tuned more for
correctness.

Co-authored-by: Ekta Khanna <ekhanna@pivotal.io>