Remove descriptions from a few GPDB operator-backing functions.
Upstream commit 28233ffaa4 instated a rule that the function backing
an operator should not have a comment, or the comment should say
"implementation of XXX operator". It's not a hard rule, we could've
added exceptions to the expected output, but seems better to just
remove the function descriptions from the cases that violate it. These
descriptions were not very helpful.