Fix ANALYZE on AO tables.
I removed the support for AO tables from RelationGetNumberOfBlocks()
earlier, because it felt like a modularity violation to get a snapshot
from such a low level function, and I thought that it was never used
on AO tables. But there was this one caller in analyze_rel_internal. Teach
it to call GetAOTotalBytes() directly, which is what
RelationGetNumberOfBlocks() used to do.

Also, fix snapshot leak in UpdateMasterAosegTotals().