I changed the query sent by ANALYZE to not schema-qualify gp_dist_random.
I did that earlier, because the code to handle gp_dist_random() didn't
accept that anymore. I'm not 100% what we want to do with that in the long
term. Should we accept pg_catalog.gp_dist_random()? What about
anyrandomschema.gp_dist_random()? Are there backwards-compatibility issues?
But in any case let's keep the expected output in sync of the queries
actually used for now.

NOTE: Should work on this in a separate PR, I don't think it ever made
sense to accept any random schema prefix for gp_dist_random.