Fix scankey initialization for ALTER TABLE ADD PARTITION.
I changed this scan from an index_begin scan to systable_beginscan_ordered
earlier, while fixing SnapshotNow issues. But with that change, the attribute
numbers in the scan key must also be changed to be attribute numbers of the
table, not the index.

This fixes the "column is not in index" errrors.