Fix GIN support for complex array.
The _complex_ops operator class for GIN indexes was incomplete. The
support functions, in pg_amproc, were missing, and the datatype in
pg_opclass was incorrect. (It was 'numeric', instead of 'complex',
a copy-pasto from _numeric_ops). As a result, trying to create a GIN
index on complex[] did not work:

postgres=# CREATE INDEX ON complex_array_tab USING gin (complex_arr);
ERROR:  missing support function 1 for attribute 1 of index "complex_array_tab_complex_arr_idx" (indexam.c:832)

This was not a problem in GPDB 5, because we didn't support GIN indexes
at all. But now that we do, let's fix it. I don't think anyone cares much
about this particular case, but we support GIN indexes on arrays of all
built-in types, so let's support this too completeness.

(Submitted to master as https://github.com/greenplum-db/gpdb/pull/5884)