Fix orafce compilation.
Cherry-pick of upstream (orafce) commit ed8c11c8af.  This is the bare
minimum to get it compiling. There were further commits in orafce to
also eliminate the compiler warnings, but I think we should rev up orafce
as whole anyway, so I'm not digging deeper into that right now.